ingest/requirements.txt
boto3==1.9.199
botocore==1.12.199
certifi==2019.6.16
chardet==3.0.4
docutils==0.14
glob2==0.7
idna==2.8
jmespath==0.9.4
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
